Update spec helper for zuul-cloner
We assume that zuul-cloner will prepare all of the git repos needed
before running beaker-rspec in CI.  Therefore, perform no module
installation if the modules directory exists.  However, if the dir
is empty (ie, the case of a developer running beaker-rspec locally)
this will clone system-config and run install_modules.

Change-Id: I81d669513a4f15dbf8879e05401605f94b2dc523
Co-Authored-By: Monty Taylor <mordred@inaugust.com>