Move jenkins job builder into jenkins_master
The idea is that an openstackci user should be able to say
"give me an openstackci::jenkins_master" and part of that should be
to set up JJB correctly.  So we move JJB setup into jenkins_master
to make that happen.

There was a request to make these JJB param names more consistent
and less ambiguous to third party devs in change Ibc7a46a8cb364
so this change also updates jjb param names to be more sensible.

Update the zuul scheduler as part of this patch to avoid puppet duplicate
definition errors for Class['project_config'] that would result when this
patch lands.

Needed-by: Id10d63745fff43f3188e630df38c8c9ba97c3e17
Change-Id: I50d4d2dd4209250e14ceeb7ff8ddf5b6ba3e2ad4