Pin fog-google gem to known working version.
`fog-google` is a transitive dependency which has recently bumped a
minor version restricting its usage on ruby 2.0.

The gate gate-openstackci-beaker-trusty-dsvm, used by other jobs, runs
on ruby 1.9.3 and because the other gems don't have a pined version, the
minor update causes the jobs to fail.

This change pins `fog-google` to the previous minor version, which does not
have the ruby 2.0 requirement.

[1] https://github.com/fog/fog-google/commit/a66b16fa7c2373f9c8be2e80bc942ad8d13ece3f

Change-Id: I43613a0b15c72acebe254edeeecb119a3d368a4e