Add Graphite node class
This patch is a first step in refactoring the existing graphite
deployment code by adding a layer of abstraction to help would-be
consumers, as opposed to providing any additional functionality
beyond what already exists.  This work is being done as part of
the effort described here:

https://review.openstack.org/#/c/239810/2/specs/openstackci-extension.rst

The goal of the overall effort is to make OpenStack CI devops code
easier to consume by those who wish to implement all or parts of
the entire OpenStack CI infrastructure (not just the minimal
requirements for third-party CI) in their own environments, for
instance for testing internal OpenStack forks.

The end result of the refactoring is to have the classes in
puppet-openstackci define NODE types (i.e. things defined inside
"node" resources in puppet site manifests) that make up the OpenStack
CI infrastructure, as well as a single-node node definition used for
testing, minimal CI environments, etc.  This is in contrast to classes
defined in modules such as puppet-graphite which contain classes defining
APPLICATIONS (i.e. things to be deployed on nodes).

The thinking is that this refactoring will help would-be consumers by
enabling them to find all the node definitions in a single place, i.e.
puppet-openstackci, instead of having to look in three places: the
site.pp file in system-config, puppet-openstackci, and puppet-<app>.

It is generally agreed that from a technical perspective this refactoring
serves no functional purpose, so pointing that out in reviews will not
bring anything new to the discussion.  The relevant questions from the
author's perspective are:

1. Is this additional layer of abstraction useful to people who want to
   consume OpenStack CI infrastructure devops code and does this usefulness
   outweigh any potential issues caused by having an additional layer of
   abstraction?
2. Should this second layer of abstraction be added only when it does more
   than serve as a wrapper?

Regarding question #2:  It is arguably better to introduce the layer of
abstraction as soon as possible, because consumers will have the abilility
to use it right away. The alternative is for them to use puppet-graphite
directly, which is problematic because they won't know to switch to the new
layer of abstraction when it is added and because of that, they won't benefit
from any additional function provided by the layer of abstraction when and if
it is added.  It is also arguably better to be consistent--one class per node
type--even if that leads to wrapper-only code that remains perpetually so.

Depends-On: Ie67b394ae50e1647b7f0386b1a2ad10f06ee5df0

Change-Id: I125d58ad4bd3ad4176d7876355a64063f55af849