Pass Java arguments to Jenkins master
Passing Java arguments to Jenkins puppet allows to tune Jenkins
master Java. This is useful in order to set various Java attributes
like Java heap size virtual memory consumption.
This is most useful for downstream projects.

This patch also adds workaround for Jenkins' SECURITY-170
Allow to use parameters that are not defined on a job from the build
environment.
https://issues.jenkins-ci.org/browse/JENKINS-34885

Change-Id: I745a23b0f177f868d5406639bc8c42e3acacea72
Depends-On: Ice0e4ffa2fffd041d8bcc4a0b323ffae7ba2b0c8