Add text and comments to clarify potential confusions on config edits
Added text to the documentation to encourage using the pristine flow
before proceeding to customizations.  It's a good reminder.

Also hint at using master so that users comformable with git have
a logical question answered earlier. It ties into the description
of use of master later in the documentation.

Comment in the zuul config regarding where to add the nodepool
cloud.yaml config.  The existing reference isn't clear on what
format to use for the input file.

Change-Id: I03ffd64d31256175a05d331f968b19a9c0f77d6b