Add flexible-publish plugin
I would like to build rpms for AFS for Centos in a periodic job and
publish them to static.openstack.org.  However, if the files are
already built, and the job didn't produce any new ones, I would like
to just leave things alone.  I can't uncondintonally copy the output
of the job, because if the job decided there was nothing to do it will
not have created any files to copy, so I would end up overwriting the
files there with a blank directory.

Using the flexible-publish approach, I should be able to see if the
job created any files before running a scp to copy them.

jjb has had support for this since
Id72a4b01f451ae2613b24347afb6493bf7641cd1 which was first released in
0.4.2

Change-Id: If9ce31aecbcb1989234259c8c45dbdb7f47fce56