Order periodic job results by date
When someone looks at the logs for a periodic job, they are almost
certainly looking for the results of the latest run.  Order the
listing by date descending so that is right up the top, and you don't
have to search through all the entries to find the latest one.

Change-Id: Ie37343728a90a805b20b2b125e17cd3b11dc6b40