Use our custom pip provider for keyring package
The pip provider in puppet hard-codes pypi.python.org in the 'latest'
method[1]. This is a problem because it bypasses our mirrors and causes
intermittent CI failures across all the puppet modules. This patch
changes the keyring package resource to use our new 'openstack_pip'
package provider[2], which uses the pip command line tool to check for
the latest version of a package, thereby using all the configs set in
/etc/pip.conf.

This change requires downstream users to update their installation of
the puppet-pip module.

[1] https://github.com/puppetlabs/puppet/blob/3.8.6/lib/puppet/provider/package/pip.rb#L69
[2] https://review.openstack.org/#/c/302396/

Change-Id: I91d87487c358c247bd12c976f50a2d0f8d785cf5