Use openstack-infra jenkins users public ssh key
In migrating to zuul-launcher, we exposed an issue where the
acceptance testing was replacing the our actually SSH public key in
the gate.  Because zuul-launcher does multiple SSH connections to the
node, we end up getting locked out when we attempt to copy over the
logs after the test has been successful.

Change-Id: Ia3e57f645cf11064451570ea0013100666d41bbc
Signed-off-by: Paul Belanger <pabelanger@redhat.com>