Pinning JJB and Zuul for 3rd party CI
Some changes affecting the syntax and use of JJB and Zuul were made in order to
use it with Zuulv3. The objective of this patch is to expose variables and pin
the versions of these services since the recommendation for now for 3rd party
CI maintainers is to not update.

Change-Id: I289186b7a25755c9258897202f01d278d0cc5a75