Improve handling of gzipped built html/web applications
If gate jobs gzip their builds before they are uploaded or if the
builds are automatically gzipped after a while on logs.o.o, the
current configuration does not try to seek out the gzipped files
with the proper mimetypes.

This commit adds missing mimetypes for common files:
- css.gz
- js.gz
- ttf.gz (fonts)

It also adds a rewrite rule that will check if there is a .gz
equivalent for the file the user is trying to seek, if there is,
it will serve it - otherwise it's business as usual.

Change-Id: I5e5d2da0ef3c385f836507e69a686821aba925a1