Refactor zuul out of zuul schedular
With the change to move puppeting zuul up a level (I47533709a) we need
to refactor setting up the zuul scheduler without having to include
zuul.

Note: The unused variables are left in place to facilitate a puppet
transition for the co-dependent changes in system-config and
puppet-openstackci. Another change will be created to clean up the
unused variables

Change-Id: Ic21889b8d32b778d6ef92e3b09c8adc594e8b000