Allow logs(-dev).o.o to serve gzipped svg files
In https://review.openstack.org/#/c/396428/ we improved logs.o.o
to be able to serve statically generated html content/web applications
that have been gzipped.
The svg format was missed in that patch so that we can fully compress
output before uploading/storing it to logs.o.o.

We are first testing it on logs-dev.o.o and we will later apply it
to logs.o.o if it works well.

Change-Id: I0354f61f24176874a6590d32cfa06bb9cf3a02e3