Fix beaker-rspec tests
The puppet-openstackci beaker tests have been failing for all the puppet
modules since they began running on Ubuntu Xenial nodes. This was
because there was no Xenial-specific nodeset file for beaker to run and
it was simply failing upon being unable to find the file.

This patch adds the ubuntu-xenial nodeset. Further, it forgoes the
install_puppet function because it attempts to install puppet 3 from the
puppetlabs apt repo, which does not support Xenial. Because we're
no longer using the install_puppet function, we also need to set default
values for the host object so that we can access things like
host['distmoduledir']. Finally, we add the libffi-dev package since the
keyring pip package requires it to build and it does not seem to be in
our Xenial images.

Change-Id: I8cc44e78116d6a7c8406e6c03756acb63c70834f