Add parameter override to disable vhd support
[1] has added ability to disable installation of vhd-util, use it to
disable vhd package install, as it is most likely not necessary, and
might cause problems deploying diskimage-builder.

Note that I am using puppet parameter class override in hiera. This is
in my opinion not a good way to change behavior, but sometimes the
quickest. So I've added a warning as well.

[1] Icb027b52fdcb790f99e90821974582da0bf6caab

Change-Id: Ifd86d06873f70d4cf6aeeec670bb6eb307817d04