Check for presence of Zuul with Zuul v3
Zuul v3 compat layer isn't going to set a ZUUL_REF var because we aren't
fetching refs from a git repo server. Instead just check if the
ZUUL_UUID is set which is a unique id for every job run by Zuul to see
if we are running under Zuul.

Change-Id: If6c122852d796427b69fc01f41c5587155c039e9