Enforce use of UTF-8 encoding for plain text files
Fixes logs where Unicode characters were appearing as garbage because
many browsers are still defaulting to "Western" charset when
the server does not give a hint (header) about encoding.

Logs files can even mess the heuristic charset detection in browsers
so is better to tell it that they are UTF-8.

Change-Id: I46d5259c59cb8f2e1a6f65a566b2a3a344e2ac09
Story: #2004028